refactor: getDayKey, getWeekKey, getMonthKey 유틸 함수로 분리
